build_kernel: Fix commits diff version references
The Perl regex was not removing the .0 patch version, because it expected a `v` to be present, which isn't.

This failed only when building the .1 patch version of a kernel. It didn't fail when passing a kernel version, because in this case, the commits diff is not printed.

The fix has been made via a refactoring, that is, by using the conventional helper short_kernel_version().